requirements.txt
websockets>=12.0
deepgram-sdk>=3.0.0
aiohttp>=3.9.0
aiofiles>=23.0.0
pydub>=0.25.1
python-dotenv>=1.0.0
openai>=1.0.0
numpy>=1.26.0
soundfile>=0.12.1
uvloop>=0.19.0
structlog>=24.0.0

# Web API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0

# Utilities
pydantic>=2.0.0
